Upgrade Travis CI system (#2252)
This upgrades the OS in the Travis CI to Bionic and GCC version to 7. This also
fixes a bug that COMPILER_FLAGS was not correctly added at build time in gcc
tests. Somehow this bug hasn't manifested so far, but after upgrading, this
failed thread sanitizer tests because -fsanitize=thread was added only at link
time and not in build time.