Followup to workaround for minification of wasm2js mem init
Emscripten's minifier mis-minifies a couple bits in the memory
init function that's used with wasm2js when not using an external
memory init file:

https://github.com/emscripten-core/emscripten/issues/8886

Previous fix worked around the bug in one place but failed to
account for another. Have now confirmed that it works with this
change in place.

Updated test cases to match.