Make try body with multiple instructions roundtrip
Previously we didn't print an additional block when there are multiple
instructions within a `try` body, so those wast files cannot be parsed
correctly, because the wast parser assumes there are two bodies within a
`try` scope: a try body and a catch body.

We don't need to print an additional block for a `catch` body because
`(catch ...)` itself serves as a scope.