Use absolute memory addresses for emasm string indexes.
Before we used 0-based indexes which meant that we needed to modify the
code calling the emask function to replace the first argument.

Now we use the string address itself as the index/code for the emasm
constant.  This allows use code to go unmodifed as the emscripten side
will accept the memory address as the index/code.

See: https://github.com/emscripten-core/emscripten/issues/9013