PR for debugging BINARY_PASS_DEBUG (DO NOT MERGE)
This is a buggy version of #2451, which lacks a few fixes. This is not
intended for being merged or reviewed. In this current state, on the
attached file a.wasm,
```
wasm-opt -all -O4 a.wasm
```
fails with a validation error, which is fixed in #2451. But
```
BINARYEN_PASS_DEBUG=1 wasm-opt -all -O4 a.wasm
```
does NOT fail, which is weird and concerning. There are cases that
something only fails when `BINARYEN_PASS_DEBUG` is on because it
performs a few more checks, but the other way around should not happen.

@kripken suggested I should upload it that he can take a look, so :)
Thanks!