Don't attempt to de-duplicate asm consts
Since we switched the using memory addresses for asm const indexes
and stopping trying to modify the code we can no loner de-duplicate
the asm const strings here.

If we want to de-dupliate in the strings in emscripten we could do that
but it seems like a marginal benefit.

This fixes the test_html5_gamepad failures which is currently showing
up on the emscripten waterfall.