Optimize passive segments in memory-packing
When memory is packed and there are passive segments, bulk memory
operations that reference those segments by index need to be updated
to reflect the new indices and possibly split into multiple
instructions that reference multiple split segments.

This PR rewrites the memory packing pass to compute and record the
transformation scheme for each of the original segments then traverse
the module updating bulk memory operations according to these schemes.