[NFC] Enforce use of `Type::` on type names
As the number of types and the length of their names grow, it becomes
less and less desirable to have them in the global wasm
namespace. This PR removes the global constexprs that mirror the value
type enum elements inside the `Type` class to enforce the use of the
`Type::` prefix everywhere the value types are used directly.