Reuse BINARYEN_API for Emscripten builds (#2435)
This is an alternative to #2361 in that it only implements reusing BINARYEN_API
so we don't have to list all the functions in build-js.sh. Differs in that it
keeps the sh file relatively straight forward without going overboard with bash
functionality. Also adds various quotes in case of whitespace in paths and makes
it so that *.sh files always use LF line endings to ease Windows support. For
instance, I am pulling the repository in Windows but compile in WSL, which, if
Git isn't properly configured to check out line endings as-is, would otherwise
break the sh files.

Fixes #2361.