Multivalue type creation and inspection
Adds the ability to create multivalue types from vectors of Type::ValueType.
Only concrete value types are allowed in multivalue type tuples. All types are
transparently interned, so their representation is still a single
uint32_t. Types can be extracted into vectors of ValueTypes, and all the value
types expand into vectors containing themselves.

Multivalue types are not yet used in the IR, but the creation and inspection
functionality is exposed in the C and JS APIs and used in tests.