Remove FunctionType from Event
This is the start of a larger refactoring to remove FunctionType
entirely and store types and signatures directly on the entities that
use them. This change forces the BinaryWriter to traverse the module
and collect types, since they are no longer stored at the module
level. While we are collecting types, we also sort them by frequency
as an optimization.