Collect all object files from the object libraries in a CMake variable (#2477)
using the `$<TARGET_OBJECTS:objlib>` syntax.  Use this variable when
adding `libbinaryen` as static or shared library.  Additionally, use the
variable with the object files to simplify the `TARGET_LINK_LIBRARIES`
commands: add the object libraries to the sources of executables and
drop the use of our libraries in `TARGET_LINK_LIBRARIES`.  (Object
libraries cannot be linked but must be used as sources.  See
https://cmake.org/pipermail/cmake/2018-June/067721.html)