Add BYN_DEBUG/BYN_TRACE macros similar to LLVM's debug system (#2496)
This allows for debug trace message to be split my channel.  So you
can pass `--debug` to simply debug everything, or `--debug=opt`
to only debug wasm-opt.

This change is the initial introduction but as a followup I hope to
convert all tracing over to this new system so we can more easily
control the debug output.