Don't include `$` with names unless outputting to wat format
The `$` is not actually part of the name, its the marker that starts
a name in the wat format.  It can be confusing to see it show up when
doing `cerr << name`, for example.

This change has Print.cpp add the `$` which seem like the right place
to do this.  Plus it revealed a bunch of places where were not calling
printName to escape all the names we were printing.