Fix renaming in FixInvokeFunctionNamesWalker (#2513)
This fixes https://github.com/emscripten-core/emscripten/issues/9950.
The issue only shows up when debug names are not present so most of
the changes in CL come from disabling debug names in the lld tests.
We want to make sure that wasm-emscripten-finalize runs fine without
debug names so I think it makes most sense to test in this mode.

The actual bugfix is in wasm-emscripten.cpp as part of the
FixInvokeFunctionNamesWalker.  The problem was the name of the function
rather than is import name was being added to importRenames.  This means
that when debug names were present (and the two names were the same)
we didn't see the bug.