Binary format code section offset tracking (#2515)
Optionally track the binary format code section offsets,
that is, when loading a binary, remember where each IR
node was read from. This is necessary for DWARF
debug info, as these are the offsets DWARF refers to.

(Note that eventually we may want to do something
else, like first read the DWARF and only then add
debug info annotations into the IR in a more LLVM-like
manner, but this is more straightforward and should be
enough to update debug lines and ranges).

This tracking adds noticeable overhead - every single
IR node adds an entry in a map - so avoid it unless
actually necessary. Specifically, if the user passes in
-g and there are actually DWARF sections in the
binary, and we are not about to remove those sections,
then we need it.

Print binary format code section offsets in text, when
printing with -g. This will help debug and test dwarf
support. It looks like

;; code offset: 0x7

as an annotation right before each node.

Also add support for -g in wasm-opt tests (unlike
a pass, it has just one - as a prefix).

Helps #2400