Add EH support for fuzzer
This adds exception handling instruction generation support for the
fuzzer. One thing this has to handle is, EH instructions do not support
Flatten and dependent passes (several passes including ReReloop and
LocalCSE), so they have to be disabled in the fuzzer when EH is enabled.
Also this adds `--disable-flatten` option to wasm-reduce, which makes
the reducer do not use Flatten and dependent passes when reducing.

Instruction generation is straightforward except for `catch`. Due to the
restriction that `exnref.pop` follows right after `catch`, we cannot do
the usual top-down generation and have to choose among several common
scenarios that start with `exnref.pop`. Also we need to exclude the
'dangling pop' expression, which contains `exnref.pop` from certain
churns in the late stage of fuzzing, in order to prevent it from being
separated from `catch`.