Don't generate redundant unreachables in Flatten
This is to remove a code pattern that generates an additional
`unreachable` after an unreachable expression. This happens when the
unreachable expression is a direct child of a block/loop/try.
Block/loop/try preserves its childrens' preludes within it, which means
childrens' preludes are not gonna escape the structure. So this kind of
code pattern is often generated by Flatten:
```wast
(block
  (some unreachable expression)
  (unreachable) ;; unnecessary
)
```
This PR removes the unnecessary `unreachable`s generated.

On the other hand, `if` does not satisfy that property because `if`'s
condition's prelude can escape `if`.

This is not for optimization (Flatten is not for optimization per se;
it's more of an enabler of other optimization passes, so we don't need
to optimize the code generated by Flatten because we have other passes
after this); This is done in order to make for #2567 to work. We are
planning to disable Flatten for exceptions now, but that's because of
`br_on_exn`; try handling part (#2567) can land now. Current Flatten
generates unnecessary `block`s because it generates unnecessary
`unreachable`s and wrap the original unreachable expression and the new
`unreachable` instruction within a `block`, ending up generating an
extra `block`s within `catch`, which is a problem, because this can
violate the invariant that `exnref.pop` should follow after `catch`.