Use TYPE_SWITCH macro for types (NFC)
This adds `TYPE_SWITCH` macro that makes sure the given type is
converted to `enum Type::ValueType`. Using this over the regular
switch-case when switch-casing over types has an advantage that when new
types are added and not handled, they will generate compile time errors.

This is another attempt at solving #2572. We have #2577, and while I'm
OK with that, and this is just to see what other people might think.

This PR is at this stage to see what people think, and I only converted
a handful of switch-cases into `TYPE_SWITCH` for now for demonstration.