Verify --version output matches CHANGELOG
The new version string looks like this:

  wasm-opt version 90 (version_90-18-g77329439d)

The version reported here is the version from the CMakeLists.txt file
followed by the git version in brackets.  We verify that the main
version here matches the CHANGELOG to prevent people from changing
one without changeing the other.

This will help with emscripten that wants to be able to programaticaly
check the --version of binaryen tools.

See https://github.com/emscripten-core/emscripten/issues/10175